Fix: Vue fix implementation, lint, and build
