+++ Vue/src/data.ts
