--- Vue/src/orig_App.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import { RouterView } from 'vue-router';
-</script>
-
-<template>
-  <div class="main">
-    <RouterView/>
-  </div>
-</template>
