--- Vue/src/views/orig_HomeView.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import HomeContent from '../components/HomeContent.vue';
-</script>
-
-<template>
-  <HomeContent text="Count"/>
-</template>
-
